style: move styled components to styles
