+++ src/Components/dokiHome/Doki.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { View } from "react-native";
 import Sprite from "../Sprite";
