+++ src/Components/styles.js
@@ -1,6 +1,6 @@
 // Index of Styled Components
 import styled from 'styled-components';
-import { View, Image } from "react-native";
+import { View, Image, ImageBackground } from "react-native";
 
 // Sprite Component
 const StyledSpriteContainer = styled(View)`
@@ -24,9 +24,31 @@ const StyledSpriteImage = styled(Image)`
   left: -${({ left }) => left}px;
 `;
 
+// DokiHome Component
+const StyledDokiHomeBackground = styled(ImageBackground)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+`;
+
+const StyledDokiEggContainer = styled(View)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 50px;
+  width: 300px;
+  height: 300px;
+`;
+
 export {
-// Sprite Component
+  // Sprite Component
   StyledSpriteContainer,
   StyledTileContainer,
   StyledSpriteImage,
+  // DokiHome component
+  StyledDokiHomeBackground,
+  StyledDokiEggContainer,
 };
